js follow, comment

+++ app/controllers/follows_controller.rb
@@ -4,23 +4,31 @@ class FollowsController < ApplicationController
   def create
     follow = Follow.new(follow_params)
     follow.user = current_user
+
     if follow.save
       StreamRails.feed_manager.follow_user(follow.user_id, follow.followable_id)
     end
-    session[:return_to] ||= request.referer
-    redirect_to session.delete(:return_to)
+
+    render 'toggle_follow', locals: {
+              target: follow, 
+              text: JSON.parse( params[:text] ), 
+              classes: params[:classes] } 
   end
 
   def destroy
     follow = Follow.find_by_id(params[:id])
-    redirect_back(fallback_location: root_path) and return unless follow
+
+    return unless follow
 
     if follow.user_id == current_user.id
       follow.destroy!
       StreamRails.feed_manager.unfollow_user(follow.user_id, follow.followable_id)
     end
-    session[:return_to] ||= request.referer
-    redirect_to session.delete(:return_to)
+
+    render 'toggle_follow', locals: {
+              target: follow, 
+              text: JSON.parse( params[:text] ), 
+              classes: params[:classes] } 
   end
 
   private
